refactor: simplified structure by removing unnecessary container component
relates #285
